Some changes in hammer

+++ hammer/hammer_mathlib.cpp
@@ -223,28 +223,28 @@ bool IsLineInside(const Vector2D &pt1, const Vector2D &pt2, int x1, int y1, int 
 
 	if((ly1 > y1) != (ly2 > y1))
 	{
-		i = lx1 + (int)((long)(y1 - ly1) * (long)(lx2 - lx1) / (long)(ly2 - ly1));
+		i = lx1 + (int)((y1 - ly1) * (lx2 - lx1) / (ly2 - ly1));
 		if(i >= x1 && i <= x2)
 			return true; // the line crosses the y1 side (left)
 	}
 
 	if((ly1 > y2) != (ly2 > y2))
 	{
-		i = lx1 + (int)((long)(y2 - ly1) * (long)(lx2 - lx1) / (long)(ly2 - ly1));
+		i = lx1 + (int)((y2 - ly1) * (lx2 - lx1) / (ly2 - ly1));
 		if(i >= x1 && i <= x2)
 			return true; // the line crosses the y2 side (right)
 	}
 
 	if((lx1 > x1) != (lx2 > x1))
 	{
-		i = ly1 + (int)((long)(x1 - lx1) * (long)(ly2 - ly1) / (long)(lx2 - lx1));
+		i = ly1 + (int)((x1 - lx1) * (ly2 - ly1) / (lx2 - lx1));
 		if(i >= y1 && i <= y2)
 			return true; // the line crosses the x1 side (down)
 	}
 
 	if((lx1 > x2) != (lx2 > x2))
 	{
-		i = ly1 + (int)((long)(x2 - lx1) * (long)(ly2 - ly1) / (long)(lx2 - lx1));
+		i = ly1 + (int)((x2 - lx1) * (ly2 - ly1) / (lx2 - lx1));
 		if(i >= y1 && i <= y2)
 			return true; // the line crosses the x2 side (up)
 	}
